src/bedrock-agentcore-itsm/requirements.txt
# Core dependencies for Bedrock AgentCore implementation
boto3>=1.34.0
botocore>=1.34.0

# Strands framework for AWS-native agent development
strands-agents>=1.0.0

# HTTP client for API Gateway integration
requests>=2.31.0
urllib3>=2.0.0

# JSON handling and data validation
pydantic>=2.0.0

# Logging and configuration
python-dotenv>=1.0.0strands-agents>=1.0.0
